requirements.txt
jinja2
fastapi
uvicorn[standard]
requests
python-telegram-bot
aiorcon
psycopg2
pysrcds
docker
numpy
aiofiles
python-multipart
fastapi-cache2[redis]
python-jose[cryptography]
passlib[bcrypt]
redis
peewee